Add alt tag for GOV UK Crown
While submitting the MOT Reminders service for an accessibility audit,
we found that the GOV UK Crown in the email notifications  was missing
an alt tag. This was flagged up during the audit, and it was recommended
that we add the alt tag to reduce confusion for users using screen
readers.