Fix failure to process files with over 50,000 rows
We convert each row of a file into an instance of `Row`. This takes some
time, so if we kept doing it for arbitrarily large files the request
would time out.

For this reason we have a limit of 50,000 rows in a file. Once we are
past 50,000 rows we stop converting each row to an instance of `Row` and
return `None` instead. This is good because it avoids all the slow code
paths while still producing a list that has the right number of items in
it (ie corresponding to the number of lines in the file).

The problem was that some of the error messaging code was looking at
these `None` rows but expecting `Row` instances. Because
`None.has_bad_recipient` is not a thing an exception was being raised.

To fix this problem this commit adds:
- a check for `None` rows in the error messaging code
- a test that explicitly looks at a file bigger than the limit (which
  we didn’t have before and would have caught this when this code was
  refactored in bfc47de1394af876be023cf6587b3c5b0db07f1f)