Force paras to stay together at bottom of page
Seeing some weird behaviour where it’s prematurely bumping the last
lines of a paragraph onto the next page.

This commit is a specific hack to stop this happening as a result of
the other changes in this PR. It seems to fix the one live letter
affected without adverse effect on the others.