save the doc dl email link route in utils
since it's used in both download and frontend, it's good to have it
defined here so it can be used by both services to take advantage
of url_for.

I've left this as a blueprint here so that it's easy to add routes
to it in the frontend for different parts of the user flow.