Preview content of letter for screenreader users
At the moment we just preview letters as an image. This is not good for
screenreader users because they have no way of reading the text embedded
in the image.

This commit adds a text preview of the letter to any page which is
showing an image of the letter. This text is visually hidden but will
still be accessible to screenreader users.

Required some refactoring of the relevant `Template` subclasses to make
reuse possible.