Don’t do smart quote replace on email addresses
first.o'last@example.com is a valid email address.

first.o’last@example.com is not a valid email address.

The code we have that does automatic smart quotes is being too greedy,
and replaces the former with the latter. This means that anyone copying
an email address from the body of an email will be copying an invalid
email address.

This commit adds some extra formatting to stop that from happening.