Recognise unicode bullets as list item signifiers
Looks like people are copy-pasting content from Word or similar into
their Notify templates.

When this is happening they’re getting unicode bullet points, which our
markdown parser doesn’t recognise as the start of a list item. So they
get something that looks a bit like a list, but without the proper
spacing or indentation.

This commit amends the regular expressions used to detect lists to also
recognise bullet characters (and makes the tests a bit more
comprehensive in the process).