fix slow regex for removing smart quotes
unbound regexes perform badly with long strings. The email regex looks
for any amount of characters, followed by an @ sign, followed by
something that looks like a domain. It starts with the first character,
looks for an @ sign by scanning characters - and if it doesn't find
one, says "guess that first character isn't the beginning of an email
address, lets try with the second character". This is O(n^2) and if
there are thousands of characters in the string (that look like the
local part of a single email address) then it'll take minutes of time.

This commit changes this so that the regex specifically expects the
string to start with either a whitespace or the start of a string. This
will capture slightly more, eg, any preceeding brackets, quotes, or
other punctuation, but by anchoring it to a rare character (whitespace)
rather than a common character (a-zA-Z0-9 + punctuation), there will be
less loops. It's still technically O(n^2) but the actual complexity
will be much smaller since it will only check each individual word
once.

this test takes ~35 seconds locally with the old regex.