Add preheader to HTML emails
Email clients take the first couple of sentences[1] of an email’s
content and use it to show a preview of the email in the inbox. This is
useful, but we are making it less useful because the first bit of
content in our emails is:
- GOV.UK (if the email has GOV.UK branding)
- the name of the organisation (for single identity brands)

Worse, some email clients (looking at you Outlook) display the URL of
any images as part of this preview. This is URL is not useful or
informative.

We can get some control over the text of this preview by making sure
that the first content in the `<body>` of the message is what we’d want
in the preview. So this commit insert the first 256 characters of the
content of the message before any branding. Then it adds some CSS to
hide this content from anyone looking at the full email.

This approach is based on the recommendations in
https://www.campaignmonitor.com/blog/email-marketing/2011/12/a-practical-guide-to-email-preheaders/

1. In Outlook on a wide monitor this can reach 250 characters, which is
   how I arrived at the 256 character truncation.