Add formatter for stripping line separators
We've had some email templates with the line
separator character (U+2028) in. This has meant
some users see it rendered as an inline box with
'LSEP' inside:

https://stackoverflow.com/questions/39603446/why-is-this-symbol-showing-up-on-chrome-and-not-firefox-or-edge

This strips them out at the point emails are
rendered, before markdown processing begins.

**Background**

Presumably it's from a cut and paste of content
into the Message field of an email template.

Part of the process we use to turn markdown into
HTML interprets line breaks characters as either
markers for `<p>` blocks or for `<br />` tags.

The standard our library uses for this identifies
line breaks as either a newline (U+000A) or
carriage return (U+000D):

https://github.github.com/gfm/#line-ending

This means line separators are ignored and make it
into the final emails.

**Solution**

Line separators don't appear as line breaks when
you paste content into the textarea element and
they're ignored in the transformation to HTML.

Stripping them out supports the expectation both
those things give, that they shouldn't effect the
format or sequence of text in an email template.