add first class timings to postage
a letter is submitted to notify. if it made the 5:30pm cut-off, it will
be printed on the next working day by DVLA. Then, the following Mon-Sat
day it will be in transit. The two mon-sat days after that it will be
delivered. That is if it is second class.

If it's first class, it will actually be delivered on the transit day.