Strip leading byte order mark from files
If one of these little buggers gets in a CSV file it becomes part of the
name of the first column.

To the user this means their ‘email address’ column is no longer
recognised as being called ‘email address’ (because it has a hidden
leading character in the name).

This commit silently removes these characters so stuff just works.