Move antivirus client to utils repository.
This code has been copied (and adjusted in the process) from
document-download-api repo, since now we also need to use it
in notifications-admin, so we need to have it in a place that
will be accessible for both those apps.

Next step will be to make make document-donwload-api call
the antivirus client from here and to delete duplicate code from
there.