Replace non-breaking spaces with line-break tags
The non-breaking spaces are being used here to
force the renderer to honor the dimension set on
table cells. Without content in them, they could
be collapsed in favour of giving the cells with
in the same row more space.

This was causing a problem for users of
screenreader software. For them, the non-breaking
spaces were added to the email body text so were
being announced.

This replaces them with <br /> tags which do not
add spaces to the email body text.