Add space character to presentational images alt
Most email clients take empty alt attributes on
images to indicate the image should be ignored by
assistive technology.

Windows Mail and Outlook 2016 still announce these
images, speaking the URL in full instead.

This is a hack to try and stop that behaviour.

I've tested it with the following combinations of
OS, email client program and screenreader:

- Windows 10 Mail with NVDA
- Windows 10 Mail with JAWS
- Mac OSX High Sierra, Gmail in Chrome with
  Voiceover
- Mac OSX High Sierra, Gmail in Safari with
  Voiceover
- Mac OSX High Sierra, Mail app with Voiceover
- IOS 12, Gmail app with Voiceover
- Android 6.0.1, Gmail app with Talkback

All except NVDA using Windows 10 Mail ignore the
image. NVDA announces it as 'graphic' but does not
read out the URL.