Encode full stops in tokens we put in URLs
When we send confirmation emails they have a tokenised URL in them.
Sometimes this token contains full stops, for example:
> 3nfu34h34.g34ug34ug

It looks like some email filters interpret this as a link to a file with
the extension `.g34ug34ug`. Since this is an unknown file extension they
filter out the email.

Our testing has shown that URL-encoding the full stop seems to get
around this problem. Any browser will take this URL and decode the full
stop before sending a request to our servers, so this should Just Work.