Change email validation to remove obscure whitespace from addresses
We were stripping all types of whitespace from the ends of email
addresses, but ignoring whitespace in the middle. We recently saw an
error where there was a zero-width whitespace in an email domain. This
commit changes the validation to now remove obscure whitespace characters
from the middle of email addresses, since users will not be able to see
that these are there.