Add function to check if letter can be cancelled
We need a way of telling whether a letter can be cancelled by a user or
not in both notifications-admin and notifcations-api, so this adds the a
function which can be used in both places.

As part of this work, a couple of timezone functions (and their tests) and the
letter processing time have been copied across from notifications-api so that
they can be used in multiple repos.