cast redis keys/values to allowed types
redis-py 3.x only allows bytes, strings, ints, longs, and floats.
This commit adds a `prepare_value` function, which is called before
exception handling to ensure that these errors are caught in dev
environments. Cast UUIDs to str manually. We might, in the future, add
further types that we can handle (eg datetimes), but for now leave it
to just these.