add delete-keys-by-pattern command to redis client
delete-keys-by-pattern is a redis script that will delete all keys
matching a given pattern. Access it by providing a glob-like string to
`redis_client.delete_cache_keys_by_pattern('my-pattern-*'). Pattern
rules are as follows:

* `h?llo` matches hello, hallo and hxllo
* `h*llo` matches hllo and heeeello
* `h[ae]llo` matches hello and hallo, but not hillo
* `h[^e]llo` matches hallo, hbllo, ... but not hello
* `h[a-b]llo` matches hallo and hbllo

Redis scripts are first registered with redis (the script is stored on
the redis server itself) and then executed separately, so to support
this we've had to add a scripts variable to the redis client. Don't
worry if you create a new redis client, or if the server restarts
though - the script is stored against a hash of itself (so separate
clients will use the same script), and redis-py handles the hash not
existing (it'll recreate the hash on the server).

The script itself is written in lua, and returns the amount of keys
deleted.