Don’t use pluses as list markers
By default markdown supports pluses as list markers. No-one is using
pluses this way in Notify templates. The only time our users are using
pluses at the start of the line is to indiciate a phone number with an
international dialing code (eg `+44 01234 5…`)

This commit changes the regular expression to not treat lines starting
with plusses as list items, but instead render them as normal text.