Break URLs onto a new line to stop very long lines
`word-break: break-all` tells the browser to break something wherever
possible. There is also the non-standard `word-break: break-word` which
says to only break a URL when it needs to (ie when it would otherwise
overflow the whole line).

I think that `break-all` is better because:
- it doesn’t leave such big gaps in the text, so looks better
- is the only one that in that CSS standard, so has comprehensive
  browser support (see https://caniuse.com/word-break)

This only applies to links where the text of the link is a URL. Links
where the template author is specifying the text of the link will
continue to break in the normal way (ie like other text).