Keep a list of characters that cause tech failure
We know that if one of these characters appears in the hostname that SES
will reject the email.

This causes a technical failure, which is something we should be keeping
an eye on. If there’s noise in the technical failure logs then we won’t
pay such close attention to them.

This commit changes the validation so that these email addresses will
never even get to the point of being sent.