Strip all whitespace from placeholders in RecipientCSV
We were not consistently stripping whitespace from around placeholders
in the `RecipientCSV` class. We were left stripping all placeholders
because of the `skipinitialspace=True` argument we pass to `csv.reader`
and we were also right stripping the last placeholder in the file, but
no other white space was removed.

In notifications-admin the extra whitespace doesn't appear (because
multiple white space characters only appear as one whitespace charcter
in HTML), which means that what got sent and what the user sees in admin
are different, and the `billable_units` can be confusing.