Put space between r and slash in <br /> tag
Our emails use the XHTML 1.0 strict doctype. The XHTML 1.0 strict
specification recommends having a space between the `br` and the closing
`/` of the `<br />` tag, for compatibility with existing user agents[1],
 although only the self-closing slash is required[2].

Some versions of the Alpine Mail client only recognise the tag with the
space before the slash[3]. Users of this client see multiple lines
concatenated into one in our emails, when the author of the email has
used a single linebreak.

1. https://www.w3.org/TR/xhtml1/#C_2
2. https://www.w3.org/TR/xhtml1/#h-4.6
3. http://mailman13.u.washington.edu/pipermail/alpine-info/2014-March/005718.html