Don’t include URLs in the preheader
Currently our preheader renders links the same way as a plain text email
does (i.e. by printing the URL after the link text).

This isn’t an accurate preview. The person opening the email won’t see
the URL, they’ll only see the link text.

So this commit removes the URLs from the preheader, but keeps the text
of the link.