Add validation for length of email local-part
We had validation for length of the domain part of email addresses, but
not for the local part and recently saw an error from SES stating that the
address length of a recipient was too long. Adding in validation that
the local part of the email address is not more than 64 characters (the
maximum) means that email addresses will fail validation if any part is
too long or if the whole thing is too long.