Preserve placeholder ordering in template change
Given some new template content like:
```
Hello ((first name)) ((last name))
```

It’s more natural to, in the admin app, say ‘You added ((first name))
and ((last name))’ rather than ‘You added ((last name))
and ((first name))’, or both in a non deterministic way.

This means preserving the order of the placeholders throughout the
various classes that are used to calculate the difference in
placeholders between two templates. The way we do this is by using
`OrderedDict` and `OrderedSet` where we would otherwise use native `set`
and `dict` objects. This is because `set` and `dict` don’t preserve the
order, at least in the version of Python we’re using.