Fix regex to not include pipe symbols
Pipe symbols were included in these two regexes under the
misunderstanding that they were needed to act like an `OR` and therefore
match any of the characters in the set. By default, regex will match if
any single character is found in the set.

This will stop us removing whitespace from around `|` so the string
` | ` will not be transformed to `| `.

Changing the first regex without changing the second would mean we would
transofmr `|` into `-`, which we also don't want to do when calling
`do_nice_typography`.