Remove references to class for easier subclassing
Some of the method in `Columns` make direct reference to the class
itself. This makes it harder to subclass and override methods

This is to make it possible to override `.make_key`, which at the moment
would have no effect because methods like `__init__` will still be
calling `Columns.make_key` directly.