Refactor to make use of __contains__
This is better than having to call dunder methods or assert that a
lookup matches the default of `.get`.

Can’t entirely get rid of __getitem__ in the second case by doing
`super()[key]` because it causes this exception:
```
TypeError: 'super' object is not subscriptable
```

Think that making the last one into a one-liner was a bit too gnarly
to read.