Improve matching quality based on crowdsourced data
We ran an experiment to get staff at GDS to enter the names of countries
and checked whether the mapping we gave was what they expected.

Experiment can be viewed here:
https://postage.cloudapps.digital/

Source code here:
https://github.com/alphagov/notifications-country-lookup-spike

This commit makes 4 changes based on the results:
- adds some more useful synonyms
- adds the Welsh names for countries
- formalises how we deal with ‘ended’ countries and territories (i.e.
  those that have ceased to exist but may or may not map to another
  country)
- treats all places in the UK postage zone the same

The last one is a bit of a judgement call. We were trying to be clever
and let people send letters to Jersey and Guernsey without a postcode.
This is different to how we were treating Wales, where we’d remove
‘Wales’ from the last line and require a postcode. We were doing this
because Royal Mail said they accepted mail sent to Jersey, etc. But
there were two problems with this:
- based on our hunch that a letter sent to Guernsey without a postcode
  would be deliverable
- didn’t account for how to handle smaller Channel Islands like Sark
  which weren’t in the Royal Mail’s list

So we’re going to be a bit stricter and require postcodes for these
places, for the sake of consistency and reducing ambiguity.