Add method + property for handling personalisation
We store the address in the database as personalisation, with specific
keys. So there will be a few places where we need to transform an
address block to this dictionary and vice versa. By putting it in utils
we can reduce duplication and ensure that we’re always doing it in the
same way.

This makes sure that the last line of the address always goes into the
`postcode` and `address line 7` fields. Legacy code will look at
postcode, newer code should ignore `postcode` if `address line 7` is
present.