Stop stripping pipes from addresses
This is a holdover from when we used to give DVLA a pipe-delimited file
of data, rather than a PDF. Because the data was pipe-delimited the
fields themselves couldn’t have pipes in.

Removing this from the address block will let us make this code simpler.
Later we can remove it from the other places too.