Let the Jinja template render list HTML
This is a less bespoke solution than the `nl2li` function. This kind of
simple templating is exactly what Jinja is designed to do.

It also handles unsanitised input, which `nl2li` didn’t do. This was
fine  before when everything was going through an instance of `Field`
becauses  that escapes HTML. But now that some addresses are coming from
the `PostalAddress` class instead we can’t assume input has been
sanitised from HTML injection.