Let any 3 lines of address render a template
Even though the API still requires address lines 1, 2 and postcode we
can get ready for when it doesn’t by letting the template-rendering bit
of the codebase render a letter with any 3 address lines.

If neither are provided it still falls back to showing all the
placeholders highlighted in yellow.

Making the decision to still render the address as valid even if the
postcode isn’t, because I think it will help people see the problem with
the postcode if the placeholders aren’t there to distract them.