Look at multiple columns to validate letter address
Previously we would look at each column of the address and see if it
had an error (not filled in, or for the postcode column, a bad
postcode).

We want to (in subsequent commits) move to not requiring specific
address lines, but a combination of any 3. So we need to combine all of
the possible ones before checking whether the address is valid. This is
what the `PostalAddress` class does, so we can use to derive errors for
a whole row when trying to send letters.