Enforce that template instances always have a type
Our CSV parsing code would be more straightforward if it could assume
that a template instance will always have a type.

Once we are requiring the type to always be passed in we can check that
it matches the kind of template object you’re trying to create. This
gives use safety in that it will be impossible to use an
`SMSMessageTemplate` to render an email, for example.