Move some stuff into the base letter template
Stuff like the address block isn’t specific to the
`LetterPreviewTemplate` – it’s only in there so other templates can
inherit from it.

But the fact that print inherits from preview is arbitrary – it could
easily work the other way round.

It’s more logical for them both to inherit from the same base class (see
previous commit) and then the stuff that they implement can be specific
just to printing or previewing.

So this commit moves the stuff that isn’t specific to printing or
previewing into the base class.