Make subject a mixin
Letter and email templates now inherit from base templates. They could
inherit from the WithSubjectTemplate, but then you have to know that
the WithSubjectTemplate itself inherits from Template.

It feels more explicit to make a mixin with just deals with subjects,
and define the base email and letter templates as just being templates,
with a subject. Which is what the naming of `WithSubjectTemplate` was
trying to do, but codifying is better than naming.