Add property for multiple cell errors
Some errors in a row are not specific to an individual cell, but result
from assembling the contents of a cell into a message.

In the future this category of error will include bad postal addresses.

This commit adds a property we can query to display these errors
appropriately in the admin app.