Add content_replaced property
Sometimes the API needs to get the content of a template with the
placeholders replaced, but not rendered into some other form like HTML.
It does this at the moment by calling `WithSubjectTemplate.__str__`
which:
- is gross
- is no longer available because `WithSubjectTemplate` has now become a
  mixin that doesn’t implement `__str__`

Instead this commit adds a property that the API can use just for this
purpose. It preserves the current functionality, ie it won’t change the
response that we’re giving to clients.

Text messages are a special case because the API returns the rendered
message exactly, rather than some other representation of its content.