Allow any 3 columns
We no longer care which of the address columns the user puts their data
into. This means they can leave `address column 1` completely empty, for
example.

But we’re still requiring them to provide address columns, 1, 2 and
postcode, even if they don’t fill them in.

This commit changes the logic for letters to require that they have
any 3 of the address columns, not a specific 3. This will require
updating the guidance and error messaging on the admin app.