Let valid addresses end with a country, as well as a postcode
When we want to start sending international letters they need to have a
country as the last line. At the moment a valid letter is one that has a
postcode as the last line.

So this commit adds some extra logic so that, if international addresses
should be allowed, a valid country is also a valid last line.

It also adds an extra property (`has_valid_last_line`) that consuming
apps can use to provide good error messages.