Bump max fragments to 6
We think this is fine to do at the network level with no problem.

Provider wise, one of our providers has a limit at 1100,
which does add a small risk that a message of 6 fragments plus a sender
name of 183 characters would run past that 1100 limit. This risk is
acknowledged and considered minimal.

6 is being chosen as the limit because we have a user who needs to send
an SMS of 780 characters and this is the lowest number of fragments to
support that user.