Fix bug accidentally allowing countries
If you ended your address with a country we would ignore the last line
when validating. This is to catch the scenario where the country is
‘United Kingdom’ or similar. We’d then look at the penultimate line to
see if it’s a postcode. And we wouldn’t include ‘United Kingdom’ in the
address, because it’s not needed to deliver the letter.

But if the country is elsewhere in the world then we shouldn’t
start looking at the penultimate line – we should straight away say that
the letter doesn’t have a UK postcode. And we shouldn’t let you send
that letter if you don’t have permission to send letters
internationally.