Add Europe and rest of world to allowed postage
Since we’re going to be able to send letters internationally now, we
need a way to style the postage for international letters differently
to first or second class letters. So this commit adds the class names
that the CSS will need to do that.

We won’t be specific about Europe vs rest of world, because our prices
are the same.