Remove functions that strip or reformat DVLA markup
DVLA had some custom markup we used to use when sending them letters as
data. It let us add headings, paragaphs, etc. So we had to make sure
that if people included this markup in their templates that it didn’t
break things.

We no longer send DVLA letters in this way, so we no longer need to
account for this.