Make postage a required parameter
I imagine that this function defaulted to second because at one point we
only had second class postage. Now that we’re adding more kinds of
postage it’s more risky to assume ‘second’. So lets make specifying the
postage mandatory when calling this function (using the * syntax, which
means required keyword arguments).