Remove whitespace around H1 in generated PDFs
There is a bug in the library we use to render PDFs of letters from
HTML. It is to do with how the width of the line of text is calculated.
This means when the text should fit on one line, it instead breaks after
the first word, in an example given by one of our users.

We’ve only seen this occur in the one example, and making even slight
alterations to the example text make the bug go away. Nevertheless, we
should ensure as best we can that our letters render as expected.

There appears to be a workaround, which is removing the trailing
whitespace from the contents of the HTML tag. This makes sense if, in
some parts of the calculation it’s including this whitespace and in
others it’s not.

It looks like a fix for the issue will be released by the library, but
for now this workaround gets our letters looking right again.