Remove obscure whitespace from postcodes
We will say postcodes are valid if they contain normal space characters,
but invalid if they contain obscure space characters. But it’s very hard
for our users to spot the difference between a non-breaking space and a
normal space, so they can’t figure out why it’s saying their postcode is
invalid.

Instead, normalising the postcode should remove all whitespace, not just
normal, ASCII 32 space characters.

It also happens that the character which unearthed this bug (a
non-breaking space) is neither in our list of obscure whitespace or
Python’s `string.whitespace`. So this commit also adds that character to
the list.