Handle empty lists as personalisation values
At the moment if you pass in an empty list the original placeholder
remains in the message:
```html
<span class='placeholder'>((content))</span>
```

However passing in an empty string gives an empty string in the
outputted message.

This commit changes the list behaviour to work like the string
behaviour. It also filters out strings that are just whitespace so that
the outputted lists never contain empty items.