Add Redis cache from admin app
Adding the admin app’s Redis cache to utils will make it possible to
reuse it in the API.

The Redis cache provides a suite of decorators which can be used on
any function that returns a JSON-serialisable type.

When the function is called with the same arguments, the value will be
fetched from Redis, rather than making an expensive database call, or
HTTP request, or whatever the underlying method is doing.

The cache key can be defined f-string style, with the available variable
names being those in the original function’s signature.

Moving this code to utils required refactoring it into a class, so that
it can be instantiated with an instance of a Redis client. The admin app
relied on importing this client from `app`, which is not possible in
utils.