Add test for processing a big CSV of addresses
So that we can benchmark how fast a spreadsheet full of addresses can be
processed, this commit adds a test which runs against 50,000 rows of
data.

The address data is randomised such that:
- there are 129,600 possible postcodes
- there are 129,600,000 possible addresses

This should prevent us from getting false positive optimisations. For
example if every address had the same postcode and the postcode
validation function was cached then it would be very fast, but only for
the test case and not real world data.