Compute frequently used properties in constructor of PostalAddress
The code within the `PostalAddress` class looks at the lines of the
address multiple times when checking its validity. Each time it does
this, the lines are re-computed for the raw address.

Computing the lines from the raw address is relatively slow, because it
involves mutating strings.

Similarly, working out the country for an address is expensive because
we need to take what might be an country, for example ‘South Africa’ and
normalise it to ‘southafrica’ in order to look it up in our dictionary
of countries.

This commit improves the performance of the `PostalAddress` class by
computing these properties when the class is instantiated. This means
that, for a given instance, they will only be computed once. Previously,
when checking an address was valid, we were splitting the raw address
into a list of lines 4 times.