Put LRU cache on postcode detection and formatting
When validating an address we look at the postcode multiple times. This
means running the same regex and string manipulation code multiple
times.

A least recently used (LRU) cache stores the result of a function call
against its arguments, so if it’s called with the same argument the code
doesn’t get re-run, but a cached value is returned.

This is preferable to computing the postcode up front, since for
international letters we never need to look at the postcode, therefore
to compute it would be redundant.