Optimise country lookup when given a postcode
We check to see if the last line of the address is a country when
instantiating a postal address. We need to do this first, because if the
last line is a synonym for the UK then we need to remove it before we
can look for a postcode.

In reality, this means that we’re often running code which says ‘is
SW1A 1AA a country?’.

One difference between postcodes and countries is that postcodes always
contain a number[1], and countries never do. And this is faster to
compute than normalising the postcode and looking it up in our
dictionary of countries.

So this commit adds that optimisation. It will make it slightly slower
to validate an international address, but that’s offset by the country
lookup being faster than the postcode regex anyway.

***

1. We have some country synonyms which do contain numbers and need to
   be removed for this optimisation to be allowed. I think this is OK
   because:
   - UM67 to UM95 are the unicode code points for the flags of the
     United States Minor Outlying islands, and are not used to refer to
     these territories in any other context
   - россия1 (Russia-1) seems to be the name of a TV station