Only compute Country instance for UK once
At the moment we set the `country` property of an address to an instance
of `Country`. This means it has useful properties like `postal_zone`.

For domestic letters the country will always be the UK. But we were
setting this by instantiating a new instance of `Country` for every
instance of `PostalAddress`.

Instantiating an instance of `Country` meant normalising the key with
which it is called, and this key has to be normalised. Which involves
some slghtly expensive string manipulation.

We can avoid this if we create and instance of `Country` for the UK at
the module level, not the instance level.