Reduce number of lines in big address CSV test
We had 50,000 lines because this is the biggest CSV file we accept.
However this means the test runs too slowly to be an acceptable unit
test. But it’s a useful test so let’s not delete it entirely. Cutting
it down to 1,000 rows means it runs in less than a second, and it’s
written in such a way that it’s easy to bump it back up when we need to
specifically investigate the speed of processing big address CSVs.