Only count characters of text message once
When we validate a spreadsheet for sending text messages we check to see
whether:
- the resulting message would be empty
- the resulting message would be longer than the maximum size of a text
  message

This means that we count the number of characters in the message twice.
Which means that we render the message twice. Which is quite costly. So
this commit changes the code to only count the content once, then cache
that value.

This takes the time to process 50,000 rows with a template that has a
single placeholder down from about 13s to about 9s.

This is a bit fiddly because we often instantiate the template once,
then change its values for each row of the spreadsheet. So we can’t just
cache the value forever once it’s been computed once.