Don’t encode text message content to count it
Encoding the text message content just means swapping one character for
another. So `len(sms_encode(content)) == len(content)`. Therefore
running `sms_encode` is pointless if all you want to know is how many
characters you have (rather than which characters).

And it turns out that encoding the content is expensive, especially for
long templates.

So if we don’t encode the content we can save a lot of time. With 50,000
rows and a template of 600 characters this reduces the time for checking
it for errors from 28s to 11s.