reduce error log to warning log for statsd errors
we send lots of statsd messages. If there's a problem with statsd, it's
likely to be a high volume at once, which means we're likely to trip
the pagerduty P1 error threshold

we see lots of statsd errors when cells roll. This is probably because
there's just another moving cog - the internal DNS server, hosted on
each cell. cells roll often and without our knowledge

there's nothing we can do about these errors, and they resolve
themselves within seconds

there's a risk we may not see these messages at all if they're reduced
to warning level, however, we can look at a tool like sentry if this is
something we're concerned about