dont send statsd data for 15 secs if DNS fails
If DNS lookup of statsd host succeeds, we'll store it for fifteen
seconds.

If the DNS lookup fails (and it's not cached) we'll raise an exception,
skip sending the statsd data, and then carry on with regular program
flow - which may well try another statsd call almost immediately.
Interestingly, if these DNS calls slow down, as we run in eventlets, we
may end up with many concurrent DNS requests, compounding any problems
the DNS server is experiencing.

To solve this, if the DNS lookup fails, cache that failed value for ~15
seconds. Then, if we encounter that failed value (None) when trying to
send, we can just quietly and quickly skip sending statsd. After ~15
seconds the cache will expire and we'll try hitting DNS again.