Add a new type of template (broadcast)
This is part of the work to add a new channel of messaging to Notify.

As far as we know at the moment broadcast messages are the same as text
messages. But when we do find differences, for example in message
length, this code gives us a place to define them.