Remove sending statsd metrics as part of the end of any request
We want to remove statsd from our main endpoints to create notifications
(see https://github.com/alphagov/notifications-api/pull/2901).

This means we need to remove the statsd metrics included in the after
request part of the request.

We won't miss these for the API, we already get these metrics from the
notify-metric-exporter.

We will in theory miss these for the admin app, however we don't actually graph
them anywhere from the admin app so we aren't actually going to miss
them for that either. We can add the gds metrics python package to apps
that we want flask route level metrics.