Optimise checking for empty emails and letters
At the moment we render an entire email or letter for every row in a
CSV file, just to check whether the content has more than zero
characters.

We have to do this in case all of the message is contained in a
placeholder, for templates that contain only `((content))`.

However rendering templates is quite expensive, because it involves a
lot of regex and string mutation. For any case where the template has
content before or after any placeholders we can skip this check –
knowing that the template has at least some content is enough.

This assumes that templates won’t have leading or trailing whitespace,
which the admin app guarantees at the point of creation.