Add serialised models from API and admin
We’re using the concept of a serialised model in both the admin and API
codebases now to mean a Python object which is derived from certain
fields of a dictionary, typically one that comes from serialising a
database object.

We’re going to use the same concept to model broadcast areas, which will
(initially) be derived from blobs of GeoJSON. But interacting directly
with the Python dictionary representation of the GeoJSON data would be
messy, so we’ll want to wrap it in an object.

Moving this code here will allow us to re-use a lot of the same tools
that we’ve already built for doing this work in the admin and the API.